moved util funcitons to util library
